requirements/dev.txt
-r base.txt

flake8==7.3.0
black==25.11.0
isort==6.1.0
pre-commit==4.3.0
django-dotenv==1.4.2
